backend/requirements.txt
mysql-connector-python==8.2.0
Pillow==10.1.0
requests==2.31.0
gtts==2.4.0
